+++ packages/ui/src/ui/containers/ACL/ACL-types.ts
@@ -1,0 +1,13 @@
+import {
+    type ObjectPermissionRowWithExpand,
+    type PreparedApprover,
+} from '../../store/selectors/acl/acl';
+
+export type ApproverRow = PreparedApprover & {
+    aggregated_row_access_predicates?: Array<string>;
+    expanded?: boolean;
+};
+export type PermissionsRow = ObjectPermissionRowWithExpand & {
+    aggregated_row_access_predicates?: Array<string>;
+    expanded?: boolean;
+};
