+++ packages/ui/src/ui/containers/ACL/ACL.scss
@@ -16,127 +16,11 @@
         }
     }
 
-    &__object-permissions {
-        #{$acl}__permissions {
-            display: flex;
-            align-items: baseline;
-
-            .elements-text {
-                white-space: normal;
-            }
-
-            &_type_allow {
-                .icon {
-                    color: var(--success-color);
-                    margin-right: 4px;
-                    padding: 0 5px;
-                }
-            }
-
-            &_type_deny {
-                .icon {
-                    color: var(--danger-color);
-                    margin-right: 7px;
-                }
-            }
-        }
-
-        #{$acl}__subjects {
-            display: flex;
-            align-items: center;
-            justify-content: space-between;
-
-            .icon {
-                margin-left: 16px;
-            }
-        }
-    }
-
     &__acl-actions {
         flex-grow: 1;
         justify-content: end;
     }
 
-    &__row {
-        &-transparent {
-            opacity: 0.5;
-        }
-
-        &_depriving {
-            opacity: 0.5;
-            background-color: var(--danger-background);
-        }
-        &_approved,
-        &_requested {
-            opacity: 0.5;
-            background-color: var(--g-color-base-misc-light);
-        }
-    }
-
-    & &__table-item {
-        box-sizing: border-box;
-        padding-left: 4px;
-        padding-right: 4px;
-
-        &_type_expand {
-            padding-left: 2px;
-            padding-right: 2px;
-        }
-
-        &_type_subjects {
-            max-width: 320px;
-            .yt-subject-link {
-                display: flex;
-            }
-            .yc-popover {
-                display: block;
-                overflow: hidden;
-                text-overflow: ellipsis;
-            }
-        }
-
-        &_type_permissions {
-            max-width: 320px;
-        }
-
-        &_type_approve-type {
-            width: calc(270px + 160px);
-        }
-
-        &_type_inheritance-mode {
-            width: 160px;
-            max-width: 200px;
-        }
-
-        &_type_actions {
-            width: 85px;
-
-            .g-button {
-                margin-top: -4px;
-                margin-bottom: -4px;
-            }
-        }
-
-        &_type_columns {
-            width: 575px;
-            white-space: normal;
-        }
-    }
-
-    &__subject-with-tvm {
-        flex-grow: 1;
-        display: flex;
-        align-items: baseline;
-        justify-content: space-between;
-    }
-
-    &__row_unrecognized {
-        background-color: var(--g-color-base-danger-light);
-
-        &:hover {
-            --hover-background: var(--g-color-base-danger-light-hover);
-        }
-    }
 
     &__meta-item {
         margin-bottom: 20px;
@@ -154,22 +38,8 @@
         margin-left: 1ex;
     }
 
-    &__action-label {
-        text-align: center;
-        min-width: 5ex;
-        margin-right: 1ex;
-        font-weight: unset;
-        flex-shrink: 0;
-    }
-
     .data-table__row:not(:last-child) {
         border-bottom: 1px solid var(--g-color-line-generic);
-    }
-
-    &__subject {
-        &_level_1 {
-            padding-left: 28px;
-        }
     }
 
     &__toolbar {
@@ -189,16 +59,5 @@
 
     &__sync-with-col-groups {
         visibility: hidden;
-    }
-
-    &__inherited {
-        padding-right: 8px;
-        &_hidden {
-            visibility: hidden;
-        }
-    }
-
-    &__inherited-icon {
-        vertical-align: middle;
     }
 }
