+++ packages/ui/src/ui/containers/ACL/ObjectPermissionsFilters/ObjectPermissionsFilters.tsx
@@ -1,11 +1,11 @@
 import cn from 'bem-cn-lite';
 import map_ from 'lodash/map';
 import React from 'react';
-import {useDispatch, useSelector} from '../../../store/redux-hooks';
 import Filter from '../../../components/Filter/Filter';
 import Select from '../../../components/Select/Select';
 import {Toolbar} from '../../../components/WithStickyToolbar/Toolbar/Toolbar';
 import {AclMode} from '../../../constants/acl';
+import {useDispatch, useSelector} from '../../../store/redux-hooks';
 import {selectObjectPermissionsTypesList} from '../../../store/selectors/acl/acl';
 import {
     selectAclRowAccessPredicateFilter,
