+++ packages/ui/src/ui/containers/ACL/hooks/use-acl-columns/i18n/dicts.ts
@@ -1,0 +1,3 @@
+import en from './en.json';
+
+export default {en};
