+++ packages/ui/src/ui/containers/ACL/hooks/use-acl-columns/i18n/index.ts
@@ -1,0 +1,5 @@
+import {addI18Keysets} from '../../../../../i18n';
+
+import dicts from './dicts';
+
+export default addI18Keysets('yt:acl-columns', dicts);
